endpoints/model_a_endpoint/pyproject.toml
[tool.poetry]
name = "my-fastapi-app"
version = "0.1.0"
description = "FastAPI service"
authors = ["You <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.104.1"
uvicorn = {version = "0.24.0", extras = ["standard"]}
pydantic = "2.7.0"
insurance-model-a = { git = "ssh://git@github.com/Mladja99/mock_models.git", subdirectory = "models/model_a", tag = "v0.1.0" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[tool.poetry.scripts]
app = "main:app"
